refactor: split explorer page widgets
