+++ src/pages/explorer/widgets/ExplorerSidebar/ui/ExplorerSectionHeader.tsx
