+++ src/pages/explorer/widgets/ExplorerSidebar/ui/FileTreeModeTabs.tsx
