+++ src/pages/explorer/widgets/FileExplorer/lib/buildFileTree.ts
