+++ src/pages/explorer/widgets/FileExplorer/lib/getInitialCollapsedFolders.ts
