+++ src/pages/explorer/widgets/FileExplorer/ui/FolderBreadcrumbSegment.tsx
