+++ src/pages/explorer/widgets/WorkspacePanel/ui/WorkspaceFileButton.tsx
