+++ src/pages/explorer/widgets/WorkspacePanel/ui/WorkspacePanel.tsx
